# jaxrs-spec-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Metatavu - ACG Pulse dispenser credit bridge",
    "version" : "0.0.1",
    "title" : "Meta-ACG Bridge"
  },
  "host" : "acg-bridge.metatavu.io",
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json;charset=utf-8" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "paths" : {
    "/v1/system/ping" : {
      "get" : {
        "tags" : [ "System" ],
        "summary" : "Ping system",
        "description" : "Ping system",
        "operationId" : "getSystemPing",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Pong",
            "schema" : {
              "$ref" : "#/definitions/Pong"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/v1/transactions" : {
      "post" : {
        "tags" : [ "Transactions" ],
        "summary" : "Creates new transaction",
        "description" : "Creates new transaction",
        "operationId" : "createTransaction",
        "parameters" : [ {
          "in" : "body",
          "name" : "transaction",
          "description" : "Transaction to be created",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Transaction"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Created transaction",
            "schema" : {
              "$ref" : "#/definitions/Transaction"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/v1/transactions/{transactionId}" : {
      "put" : {
        "tags" : [ "Transactions" ],
        "summary" : "Updates transaction",
        "description" : "Updates transaction",
        "operationId" : "updateTransaction",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "description" : "transactionId",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Payload",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Transaction"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "The transaction were canceled successfully."
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Transactions" ],
        "summary" : "Cancel transaction",
        "description" : "Cancel transaction",
        "operationId" : "cancelTransaction",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "description" : "transactionId",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "The transaction were canceled successfully."
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Pong" : {
      "type" : "string",
      "description" : "Ping response"
    },
    "Transaction" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "description" : "Unique identifier for the transaction"
        },
        "status" : {
          "type" : "string",
          "description" : "Transaction status. One of PENDING, ERRORED, CANCELLED, SUCCESS or TIMED_OUT"
        },
        "orderId" : {
          "type" : "string",
          "description" : "Unique identifier for the orderId"
        },
        "machineId" : {
          "type" : "string",
          "description" : "Unique identifier for a machine that initiated the transaction"
        },
        "serverId" : {
          "type" : "string",
          "description" : "Unique identifier for a server that initiated the transaction"
        },
        "successUrl" : {
          "type" : "string",
          "description" : "URL address to callback when transaction is successful"
        },
        "failureUrl" : {
          "type" : "string",
          "description" : "URL address to callback when transaction has failed."
        },
        "paymentStrategy" : {
          "type" : "string",
          "description" : "Name of payment strategy used for transactions"
        },
        "amount" : {
          "type" : "number",
          "format" : "double",
          "description" : "The amount for the payment."
        },
        "properties" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/TransactionProperty"
          }
        }
      },
      "description" : "A single transaction"
    },
    "TransactionProperty" : {
      "type" : "object",
      "properties" : {
        "key" : {
          "type" : "string",
          "description" : "Key"
        },
        "value" : {
          "type" : "string",
          "description" : "Value"
        }
      },
      "description" : "A key - value pair describing a transaction property"
    },
    "InternalServerError" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "Forbidden" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "BadRequest" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "NotImplemented" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "NotFound" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    }
  }
}